requirements.txt
Flask==3.0.2
matplotlib==3.8.2
fpdf2==2.7.5
gunicorn==21.2.0